# docs/openapi.yaml
openapi: 3.1.0
info: { title: Scalora CRM API, version: 0.1.0 }
servers: [{ url: /api/v1 }]
paths:
  /health: { get: { summary: Service health, responses: { '200': { description: Healthy } } } }
  /auth/register: { post: { summary: Register owner and workspace, responses: { '201': { description: Created } } } }
  /auth/login: { post: { summary: Sign in, responses: { '200': { description: Authenticated }, '401': { description: Invalid credentials } } } }
  /auth/refresh: { post: { summary: Rotate refresh token, responses: { '200': { description: Refreshed } } } }
  /auth/me: { get: { summary: Current user and workspace, security: [{ bearerAuth: [] }], responses: { '200': { description: Current session } } } }
  /business/onboarding: { patch: { summary: Complete owner onboarding, security: [{ bearerAuth: [] }], responses: { '200': { description: Updated } } } }
  /customers: { get: { summary: List tenant customers, security: [{ bearerAuth: [] }], responses: { '200': { description: Customer page } } }, post: { summary: Create customer, security: [{ bearerAuth: [] }], responses: { '201': { description: Created } } } }
  /customers/{id}: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], get: { summary: Customer profile, security: [{ bearerAuth: [] }], responses: { '200': { description: Customer } } }, patch: { summary: Update customer, security: [{ bearerAuth: [] }], responses: { '200': { description: Updated } } }, delete: { summary: Soft-delete customer, security: [{ bearerAuth: [] }], responses: { '200': { description: Deleted } } } }
  /follow-ups: { get: { summary: List follow-ups, security: [{ bearerAuth: [] }], responses: { '200': { description: Follow-ups } } }, post: { summary: Create follow-up, security: [{ bearerAuth: [] }], responses: { '201': { description: Created } } } }
  /pipeline: { get: { summary: Get tenant pipeline, security: [{ bearerAuth: [] }], responses: { '200': { description: Pipeline } } } }
  /dashboard/summary: { get: { summary: Phase 1 metrics, security: [{ bearerAuth: [] }], responses: { '200': { description: Dashboard summary } } } }
  /conversations: { get: { summary: List tenant conversations, security: [{ bearerAuth: [] }], responses: { '200': { description: Conversation page } } }, post: { summary: Open a mock conversation for a customer, security: [{ bearerAuth: [] }], responses: { '201': { description: Conversation ready } } } }
  /conversations/{id}: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], get: { summary: Conversation messages and internal notes, security: [{ bearerAuth: [] }], responses: { '200': { description: Conversation detail } } } }
  /conversations/{id}/messages: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Send a public message, security: [{ bearerAuth: [] }], responses: { '201': { description: Message accepted } } } }
  /conversations/{id}/internal-notes: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Add an employee-only note, security: [{ bearerAuth: [] }], responses: { '201': { description: Note created } } } }
  /conversations/{id}/assignment: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], patch: { summary: Assign an active workspace member, security: [{ bearerAuth: [] }], responses: { '200': { description: Assignment updated } } } }
  /conversations/mock/incoming: { post: { summary: Simulate an inbound WhatsApp message in development, security: [{ bearerAuth: [] }], responses: { '201': { description: Message received }, '409': { description: Mock provider disabled in production } } } }
  /whatsapp/webhook: { get: { summary: Verify Meta webhook subscription, responses: { '200': { description: Challenge accepted } } }, post: { summary: Receive a signed Meta webhook event, responses: { '200': { description: Event persisted for processing }, '401': { description: Invalid signature } } } }
  /whatsapp/account: { get: { summary: Get connection metadata without credentials, security: [{ bearerAuth: [] }], responses: { '200': { description: Connection metadata } } }, delete: { summary: Disconnect WhatsApp, security: [{ bearerAuth: [] }], responses: { '200': { description: Disconnected } } } }
  /whatsapp/account/connect: { post: { summary: Connect and validate WhatsApp credentials, security: [{ bearerAuth: [] }], responses: { '201': { description: Connected } } } }
  /whatsapp/templates: { get: { summary: List synchronized templates, security: [{ bearerAuth: [] }], responses: { '200': { description: Templates } } } }
  /whatsapp/templates/sync: { post: { summary: Synchronize templates from Meta, security: [{ bearerAuth: [] }], responses: { '200': { description: Synchronized } } } }
  /whatsapp/templates/{templateId}/send: { parameters: [{ name: templateId, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Send an approved template with variables, security: [{ bearerAuth: [] }], responses: { '201': { description: Template message sent } } } }
  /campaigns: { get: { summary: List tenant campaigns and delivery counters, security: [{ bearerAuth: [] }], responses: { '200': { description: Campaign page } } }, post: { summary: Create a draft with an approved marketing template, security: [{ bearerAuth: [] }], responses: { '201': { description: Draft created } } } }
  /campaigns/estimate: { post: { summary: Estimate opted-in recipients after audience filters, security: [{ bearerAuth: [] }], responses: { '200': { description: Eligible count and sample } } } }
  /campaigns/{id}/launch: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Freeze eligible recipients and queue delivery, security: [{ bearerAuth: [] }], responses: { '200': { description: Campaign queued } } } }
  /campaigns/{id}/report: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], get: { summary: Campaign delivery, read, reply, failure, and skip rates, security: [{ bearerAuth: [] }], responses: { '200': { description: Campaign report } } } }
  /consent/customers: { get: { summary: List customers by current marketing consent, security: [{ bearerAuth: [] }], responses: { '200': { description: Consent list } } } }
  /consent/customers/{id}: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], patch: { summary: Record an opt-in or opt-out, security: [{ bearerAuth: [] }], responses: { '200': { description: Consent updated } } } }
  /products: { get: { summary: List tenant catalog items, security: [{ bearerAuth: [] }], responses: { '200': { description: Product and service catalog } } }, post: { summary: Create a product or service, security: [{ bearerAuth: [] }], responses: { '201': { description: Catalog item created } } } }
  /quotations: { get: { summary: List quotations, security: [{ bearerAuth: [] }], responses: { '200': { description: Quotations } } }, post: { summary: Create a quotation with calculated totals, security: [{ bearerAuth: [] }], responses: { '201': { description: Quotation created } } } }
  /quotations/{id}/convert: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Atomically convert a quotation to an invoice or order, security: [{ bearerAuth: [] }], responses: { '201': { description: Converted document } } } }
  /quotations/{id}/pdf: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], get: { summary: Download a printable quotation PDF, security: [{ bearerAuth: [] }], responses: { '200': { description: PDF document } } } }
  /invoices: { get: { summary: List invoices and balances, security: [{ bearerAuth: [] }], responses: { '200': { description: Invoices } } }, post: { summary: Create an invoice, security: [{ bearerAuth: [] }], responses: { '201': { description: Invoice created } } } }
  /invoices/{id}/payments: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Record a partial or full invoice payment, security: [{ bearerAuth: [] }], responses: { '201': { description: Payment recorded } } } }
  /invoices/{id}/pdf: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], get: { summary: Download a printable invoice PDF, security: [{ bearerAuth: [] }], responses: { '200': { description: PDF document } } } }
  /invoices/{id}/send-whatsapp: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Send an invoice summary using session or approved template rules, security: [{ bearerAuth: [] }], responses: { '201': { description: Message sent } } } }
  /orders: { get: { summary: List tenant orders, security: [{ bearerAuth: [] }], responses: { '200': { description: Orders } } }, post: { summary: Create an order, security: [{ bearerAuth: [] }], responses: { '201': { description: Order created } } } }
  /orders/{id}/status: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], patch: { summary: Update fulfilment status and activity history, security: [{ bearerAuth: [] }], responses: { '200': { description: Status updated } } } }
  /orders/{id}/payments: { parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }], post: { summary: Record an order payment, security: [{ bearerAuth: [] }], responses: { '201': { description: Payment recorded } } } }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }