# server/src/modules/invoices/invoice.routes.ts
import { randomUUID } from "node:crypto";import { Router } from "express";import { z } from "zod";import { prisma } from "../../lib/prisma.js";import { AppError } from "../../lib/errors.js";import { asyncHandler } from "../../lib/async-handler.js";import { ok } from "../../lib/response.js";import { authenticate,requireRole } from "../../middleware/auth.js";import { routeParam } from "../../lib/route-param.js";import { documentBaseSchema,paymentSchema } from "../commerce/document.schemas.js";import { generateDocumentPdf } from "../commerce/pdf.service.js";import { getCloudProvider } from "../whatsapp/account.service.js";import { createInvoice,recordInvoicePayment } from "./invoice.service.js";
export const invoiceRouter=Router();invoiceRouter.use(authenticate);const createSchema=documentBaseSchema.extend({issueDate:z.coerce.date().default(()=>new Date()),dueDate:z.coerce.date()});
invoiceRouter.get("/",asyncHandler(async(req,res)=>{await prisma.invoice.updateMany({where:{businessId:req.auth!.businessId,deletedAt:null,dueDate:{lt:new Date()},paymentStatus:{in:["UNPAID","PARTIALLY_PAID"]}},data:{paymentStatus:"OVERDUE"}});const query=z.object({status:z.enum(["UNPAID","PARTIALLY_PAID","PAID","OVERDUE","CANCELLED"]).optional()}).parse(req.query);const rows=await prisma.invoice.findMany({where:{businessId:req.auth!.businessId,deletedAt:null,...(query.status?{paymentStatus:query.status}:{})},include:{customer:{select:{firstName:true,lastName:true}},_count:{select:{items:true,payments:true}}},orderBy:{createdAt:"desc"}});return ok(res,rows);}));
invoiceRouter.post("/",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>ok(res,await createInvoice(req.auth!.businessId,req.auth!.userId,createSchema.parse(req.body)),"Invoice created",201)));
invoiceRouter.get("/:id",asyncHandler(async(req,res)=>{const row=await prisma.invoice.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null},include:{items:true,payments:{include:{recordedBy:{select:{firstName:true,lastName:true}}},orderBy:{paidAt:"desc"}},customer:true,sourceQuotation:{select:{id:true,quotationNumber:true}}}});if(!row)throw new AppError(404,"INVOICE_NOT_FOUND","Invoice was not found");return ok(res,row);}));
invoiceRouter.post("/:id/payments",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>ok(res,await recordInvoicePayment(req.auth!.businessId,req.auth!.userId,routeParam(req.params.id),paymentSchema.parse(req.body)),"Payment recorded",201)));
invoiceRouter.patch("/:id/cancel",requireRole("OWNER","ADMIN"),asyncHandler(async(req,res)=>{const changed=await prisma.invoice.updateMany({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null,amountPaid:0},data:{paymentStatus:"CANCELLED"}});if(!changed.count)throw new AppError(409,"INVOICE_NOT_CANCELLABLE","Paid invoices cannot be cancelled");return ok(res,null,"Invoice cancelled");}));
invoiceRouter.get("/:id/pdf",asyncHandler(async(req,res)=>{const row=await prisma.invoice.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null},include:{items:true,customer:true,business:true}});if(!row)throw new AppError(404,"INVOICE_NOT_FOUND","Invoice was not found");const pdf=await generateDocumentPdf({title:"Invoice",number:row.invoiceNumber,currency:row.currency,issueDate:row.issueDate,endDateLabel:"Due date",endDate:row.dueDate,business:{name:row.business.name,address:row.business.address,phone:row.business.phone,email:row.business.email,taxNumber:row.business.taxNumber,logoUrl:row.business.logoUrl},customer:{name:`${row.customer.firstName} ${row.customer.lastName??""}`.trim(),address:row.customer.address,phone:row.customer.phone,email:row.customer.email},items:row.items,subtotal:row.subtotal,discount:row.discount,tax:row.tax,total:row.total,amountPaid:row.amountPaid,amountDue:row.amountDue,notes:row.notes,terms:row.terms});res.setHeader("Content-Type","application/pdf");res.setHeader("Content-Disposition",`inline; filename="${row.invoiceNumber}.pdf"`);res.send(pdf);}));
invoiceRouter.post("/:id/send-whatsapp",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>{
  const input=z.object({templateId:z.string().uuid().optional(),variables:z.array(z.string().max(1024)).max(20).default([])}).parse(req.body);
  const invoice=await prisma.invoice.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null},include:{customer:true}});
  if(!invoice||!invoice.customer.normalizedPhone)throw new AppError(404,"INVOICE_CUSTOMER_NOT_FOUND","Invoice customer needs a valid phone");
  const recipientPhone=invoice.customer.normalizedPhone;
  const conversation=await prisma.conversation.upsert({where:{businessId_customerId_channel:{businessId:req.auth!.businessId,customerId:invoice.customerId,channel:"WHATSAPP"}},update:{},create:{businessId:req.auth!.businessId,customerId:invoice.customerId,channel:"WHATSAPP",status:"OPEN"}});
  const withinSession=Boolean(conversation.sessionExpiresAt&&conversation.sessionExpiresAt>new Date());
  const template=withinSession?null:input.templateId?await prisma.whatsAppTemplate.findFirst({where:{id:input.templateId,businessId:req.auth!.businessId,status:"APPROVED"}}):null;
  if(!withinSession&&(!template||!Array.isArray(template.variables)||template.variables.length!==input.variables.length))throw new AppError(409,"TEMPLATE_REQUIRED","Select an approved template and provide all variables outside the service window");
  const{account,provider}=await getCloudProvider(req.auth!.businessId);if(!withinSession&&!provider.sendTemplate)throw new AppError(500,"PROVIDER_ERROR","WhatsApp provider does not support templates");
  const body=`Invoice ${invoice.invoiceNumber}: ${invoice.currency} ${invoice.total.toString()} · Amount due ${invoice.currency} ${invoice.amountDue.toString()}`;
  const queued=await prisma.message.create({data:{businessId:req.auth!.businessId,conversationId:conversation.id,whatsAppAccountId:account.id,direction:"OUTBOUND",type:"TEXT",status:"QUEUED",body,idempotencyKey:`invoice:${invoice.id}:${randomUUID()}`,...(template?{templateId:template.id,templateName:template.name,templateCategory:template.category}:{})}});
  try{const result=withinSession?await provider.send({businessId:req.auth!.businessId,conversationId:queued.conversationId,messageId:queued.id,recipientPhone,type:"TEXT",body}):await provider.sendTemplate!({recipientPhone,templateName:template!.name,language:template!.language,bodyVariables:input.variables});const sent=await prisma.message.update({where:{id:queued.id},data:{status:"SENT",providerMessageId:result.providerMessageId,sentAt:result.acceptedAt}});return ok(res,sent,"Invoice sent through WhatsApp",201);}
  catch(error){await prisma.message.update({where:{id:queued.id},data:{status:"FAILED",errorCode:"INVOICE_SEND_FAILED",errorMessage:"Invoice message could not be sent"}});throw error;}
}));