# server/src/modules/orders/order.routes.ts
import { Router } from "express";import { z } from "zod";import { prisma } from "../../lib/prisma.js";import { AppError } from "../../lib/errors.js";import { asyncHandler } from "../../lib/async-handler.js";import { ok } from "../../lib/response.js";import { authenticate,requireRole } from "../../middleware/auth.js";import { routeParam } from "../../lib/route-param.js";import { documentBaseSchema,paymentSchema } from "../commerce/document.schemas.js";import { createOrder,recordOrderPayment } from "./order.service.js";import { updateCustomerCommerceTotals } from "../commerce/document.service.js";
export const orderRouter=Router();orderRouter.use(authenticate);const createSchema=documentBaseSchema.omit({notes:true,terms:true}).extend({deliveryFee:z.coerce.number().nonnegative().default(0),paymentMethod:z.enum(["CASH_ON_DELIVERY","CASH","CARD","BANK_TRANSFER","WHISH","OMT","OTHER"]),assignedUserId:z.string().uuid().nullable().optional(),deliveryAddress:z.string().trim().max(500).nullable().optional(),customerNotes:z.string().trim().max(2000).nullable().optional(),internalNotes:z.string().trim().max(2000).nullable().optional()});
orderRouter.get("/",asyncHandler(async(req,res)=>{const query=z.object({status:z.enum(["NEW","CONFIRMED","PREPARING","READY","OUT_FOR_DELIVERY","DELIVERED","CANCELLED","RETURNED"]).optional()}).parse(req.query);const rows=await prisma.order.findMany({where:{businessId:req.auth!.businessId,deletedAt:null,...(query.status?{status:query.status}:{}),...(req.auth!.role==="SALES_AGENT"?{assignedUserId:req.auth!.userId}:{})},include:{customer:{select:{firstName:true,lastName:true}},assignedUser:{select:{firstName:true,lastName:true}},_count:{select:{items:true,payments:true}}},orderBy:{createdAt:"desc"}});return ok(res,rows);}));
orderRouter.post("/",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>{const input=createSchema.parse(req.body);return ok(res,await createOrder(req.auth!.businessId,req.auth!.userId,{...input,assignedUserId:input.assignedUserId??(req.auth!.role==="SALES_AGENT"?req.auth!.userId:null)}),"Order created",201);}));
orderRouter.get("/:id",asyncHandler(async(req,res)=>{const row=await prisma.order.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null,...(req.auth!.role==="SALES_AGENT"?{assignedUserId:req.auth!.userId}:{})},include:{items:true,payments:{orderBy:{paidAt:"desc"}},customer:true,assignedUser:{select:{id:true,firstName:true,lastName:true}},sourceQuotation:{select:{id:true,quotationNumber:true}}}});if(!row)throw new AppError(404,"ORDER_NOT_FOUND","Order was not found");return ok(res,row);}));
orderRouter.patch("/:id/status",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>{const status=z.object({status:z.enum(["NEW","CONFIRMED","PREPARING","READY","OUT_FOR_DELIVERY","DELIVERED","CANCELLED","RETURNED"])}).parse(req.body).status;const id=routeParam(req.params.id);const order=await prisma.order.findFirst({where:{id,businessId:req.auth!.businessId,deletedAt:null,...(req.auth!.role==="SALES_AGENT"?{assignedUserId:req.auth!.userId}:{})}});if(!order)throw new AppError(404,"ORDER_NOT_FOUND","Order was not found");await prisma.$transaction(async tx=>{await tx.order.update({where:{id},data:{status,...(status==="CANCELLED"&&order.amountPaid.isZero()?{paymentStatus:"CANCELLED"}:{})}});await tx.activity.create({data:{businessId:req.auth!.businessId,customerId:order.customerId,actorId:req.auth!.userId,type:"ORDER_STATUS_CHANGED",metadata:{orderId:id,from:order.status,to:status}}});await updateCustomerCommerceTotals(tx,req.auth!.businessId,order.customerId);});return ok(res,null,"Order status updated");}));
orderRouter.post("/:id/payments",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>ok(res,await recordOrderPayment(req.auth!.businessId,req.auth!.userId,routeParam(req.params.id),paymentSchema.parse(req.body)),"Order payment recorded",201)));