# server/src/modules/quotations/quotation.routes.ts
import { Router } from "express";import { z } from "zod";import { prisma } from "../../lib/prisma.js";import { AppError } from "../../lib/errors.js";import { asyncHandler } from "../../lib/async-handler.js";import { ok } from "../../lib/response.js";import { authenticate,requireRole } from "../../middleware/auth.js";import { routeParam } from "../../lib/route-param.js";import { documentBaseSchema } from "../commerce/document.schemas.js";import { generateDocumentPdf } from "../commerce/pdf.service.js";import { convertQuotation,createQuotation } from "./quotation.service.js";
export const quotationRouter=Router();quotationRouter.use(authenticate);const createSchema=documentBaseSchema.extend({issueDate:z.coerce.date().default(()=>new Date()),expiryDate:z.coerce.date()});
quotationRouter.get("/",asyncHandler(async(req,res)=>{await prisma.quotation.updateMany({where:{businessId:req.auth!.businessId,deletedAt:null,expiryDate:{lt:new Date()},status:{in:["DRAFT","SENT","VIEWED"]}},data:{status:"EXPIRED"}});const query=z.object({status:z.enum(["DRAFT","SENT","VIEWED","ACCEPTED","REJECTED","EXPIRED","CONVERTED"]).optional()}).parse(req.query);const rows=await prisma.quotation.findMany({where:{businessId:req.auth!.businessId,deletedAt:null,...(query.status?{status:query.status}:{})},include:{customer:{select:{firstName:true,lastName:true}},_count:{select:{items:true}}},orderBy:{createdAt:"desc"}});return ok(res,rows);}));
quotationRouter.post("/",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>ok(res,await createQuotation(req.auth!.businessId,req.auth!.userId,createSchema.parse(req.body)),"Quotation created",201)));
quotationRouter.get("/:id",asyncHandler(async(req,res)=>{const row=await prisma.quotation.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null},include:{items:true,customer:true,createdBy:{select:{firstName:true,lastName:true}},convertedInvoice:{select:{id:true,invoiceNumber:true}},convertedOrder:{select:{id:true,orderNumber:true}}}});if(!row)throw new AppError(404,"QUOTATION_NOT_FOUND","Quotation was not found");return ok(res,row);}));
quotationRouter.patch("/:id/status",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>{const status=z.object({status:z.enum(["DRAFT","SENT","VIEWED","ACCEPTED","REJECTED","EXPIRED"])}).parse(req.body).status;const changed=await prisma.quotation.updateMany({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null,status:{not:"CONVERTED"}},data:{status}});if(!changed.count)throw new AppError(409,"QUOTATION_NOT_EDITABLE","Quotation status cannot be changed");return ok(res,null,"Quotation status updated");}));
quotationRouter.post("/:id/convert",requireRole("OWNER","ADMIN","SALES_AGENT"),asyncHandler(async(req,res)=>{const target=z.object({target:z.enum(["invoice","order"])}).parse(req.body).target;return ok(res,await convertQuotation(req.auth!.businessId,req.auth!.userId,routeParam(req.params.id),target),`Quotation converted to ${target}`,201);}));
quotationRouter.get("/:id/pdf",asyncHandler(async(req,res)=>{const row=await prisma.quotation.findFirst({where:{id:routeParam(req.params.id),businessId:req.auth!.businessId,deletedAt:null},include:{items:true,customer:true,business:true}});if(!row)throw new AppError(404,"QUOTATION_NOT_FOUND","Quotation was not found");const pdf=await generateDocumentPdf({title:"Quotation",number:row.quotationNumber,currency:row.currency,issueDate:row.issueDate,endDateLabel:"Expiry date",endDate:row.expiryDate,business:{name:row.business.name,address:row.business.address,phone:row.business.phone,email:row.business.email,taxNumber:row.business.taxNumber,logoUrl:row.business.logoUrl},customer:{name:`${row.customer.firstName} ${row.customer.lastName??""}`.trim(),address:row.customer.address,phone:row.customer.phone,email:row.customer.email},items:row.items,subtotal:row.subtotal,discount:row.discount,tax:row.tax,total:row.total,notes:row.notes,terms:row.terms});res.setHeader("Content-Type","application/pdf");res.setHeader("Content-Disposition",`inline; filename="${row.quotationNumber}.pdf"`);res.send(pdf);}));